# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bilingual Reader API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookTranslationStatus =
  (typeof BookTranslationStatus)[keyof typeof BookTranslationStatus];

export const BookTranslationStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface Book {
  id: number;
  title: string;
  author?: string;
  language: string;
  totalParagraphs: number;
  translatedParagraphs: number;
  translationStatus: BookTranslationStatus;
  createdAt: string;
}

export interface CreateBookBody {
  title: string;
  author?: string;
  language?: string;
  /** Full text content of the book */
  content: string;
}

export interface Paragraph {
  id: number;
  bookId: number;
  position: number;
  originalText: string;
  translatedText?: string | null;
  isTranslated: boolean;
}

export interface ParagraphsPage {
  paragraphs: Paragraph[];
  total: number;
  page: number;
  pageSize: number;
  totalPages: number;
}

export type TranslationStatusStatus =
  (typeof TranslationStatusStatus)[keyof typeof TranslationStatusStatus];

export const TranslationStatusStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface TranslationStatus {
  bookId: number;
  status: TranslationStatusStatus;
  totalParagraphs: number;
  translatedParagraphs: number;
  progressPercent: number;
}

export interface StartTranslationBody {
  /** Number of paragraphs to translate per batch */
  batchSize?: number;
}

export interface DictionaryEntry {
  word: string;
  translations: string[];
  partOfSpeech?: string;
  examples?: string[];
  lookedUpAt?: string;
}

export interface BookStats {
  bookId: number;
  totalParagraphs: number;
  translatedParagraphs: number;
  wordCount: number;
  uniqueWordsLookedUp: number;
  progressPercent: number;
}

export type ListParagraphsParams = {
  page?: number;
  pageSize?: number;
};

export type LookupWordParams = {
  word: string;
  context?: string;
};