# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bilingual Reader API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all books
 */
export const ListBooksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string().optional(),
  language: zod.string(),
  totalParagraphs: zod.number(),
  translatedParagraphs: zod.number(),
  translationStatus: zod.enum(["pending", "in_progress", "completed"]),
  createdAt: zod.coerce.date(),
});
export const ListBooksResponse = zod.array(ListBooksResponseItem);

/**
 * @summary Upload a new book
 */
export const createBookBodyLanguageDefault = `en`;

export const CreateBookBody = zod.object({
  title: zod.string(),
  author: zod.string().optional(),
  language: zod.string().default(createBookBodyLanguageDefault),
  content: zod.string().describe("Full text content of the book"),
});

/**
 * @summary Get a book by ID
 */
export const GetBookParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  author: zod.string().optional(),
  language: zod.string(),
  totalParagraphs: zod.number(),
  translatedParagraphs: zod.number(),
  translationStatus: zod.enum(["pending", "in_progress", "completed"]),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a book
 */
export const DeleteBookParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List paragraphs of a book
 */
export const ListParagraphsParams = zod.object({
  id: zod.coerce.number(),
});

export const listParagraphsQueryPageDefault = 1;
export const listParagraphsQueryPageSizeDefault = 20;

export const ListParagraphsQueryParams = zod.object({
  page: zod.coerce.number().default(listParagraphsQueryPageDefault),
  pageSize: zod.coerce.number().default(listParagraphsQueryPageSizeDefault),
});

export const ListParagraphsResponse = zod.object({
  paragraphs: zod.array(
    zod.object({
      id: zod.number(),
      bookId: zod.number(),
      position: zod.number(),
      originalText: zod.string(),
      translatedText: zod.string().nullish(),
      isTranslated: zod.boolean(),
    }),
  ),
  total: zod.number(),
  page: zod.number(),
  pageSize: zod.number(),
  totalPages: zod.number(),
});

/**
 * @summary Get translation progress for a book
 */
export const GetTranslationStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTranslationStatusResponse = zod.object({
  bookId: zod.number(),
  status: zod.enum(["pending", "in_progress", "completed"]),
  totalParagraphs: zod.number(),
  translatedParagraphs: zod.number(),
  progressPercent: zod.number(),
});

/**
 * @summary Start translating a book (SSE stream)
 */
export const StartTranslationParams = zod.object({
  id: zod.coerce.number(),
});

export const startTranslationBodyBatchSizeDefault = 10;

export const StartTranslationBody = zod.object({
  batchSize: zod
    .number()
    .default(startTranslationBodyBatchSizeDefault)
    .describe("Number of paragraphs to translate per batch"),
});

/**
 * @summary Get the translation for a specific paragraph
 */
export const GetParagraphTranslationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetParagraphTranslationResponse = zod.object({
  id: zod.number(),
  bookId: zod.number(),
  position: zod.number(),
  originalText: zod.string(),
  translatedText: zod.string().nullish(),
  isTranslated: zod.boolean(),
});

/**
 * @summary Look up a word in the English-Russian dictionary
 */
export const LookupWordQueryParams = zod.object({
  word: zod.coerce.string(),
  context: zod.coerce.string().optional(),
});

export const LookupWordResponse = zod.object({
  word: zod.string(),
  translations: zod.array(zod.string()),
  partOfSpeech: zod.string().optional(),
  examples: zod.array(zod.string()).optional(),
  lookedUpAt: zod.coerce.date().optional(),
});

/**
 * @summary Get recently looked up words
 */
export const GetRecentLookupsResponseItem = zod.object({
  word: zod.string(),
  translations: zod.array(zod.string()),
  partOfSpeech: zod.string().optional(),
  examples: zod.array(zod.string()).optional(),
  lookedUpAt: zod.coerce.date().optional(),
});
export const GetRecentLookupsResponse = zod.array(GetRecentLookupsResponseItem);

/**
 * @summary Get reading stats for a book
 */
export const GetBookStatsParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookStatsResponse = zod.object({
  bookId: zod.number(),
  totalParagraphs: zod.number(),
  translatedParagraphs: zod.number(),
  wordCount: zod.number(),
  uniqueWordsLookedUp: zod.number(),
  progressPercent: zod.number(),
});